+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - height of `textarea` elements to change dynamically with content
 
+### Added
+
+- `<MultiFactorAuthentication>` & `<AppUserMfaProvider>` components
+- `useUserMeetsMfaRequirement` hook
+
 ## [6.9.2] - 2024-11-13
 
 ### Dependencies